+++ tsconfig.json
@@ -11,7 +11,9 @@
     "./scripts",
     "./test",
     "./deploy",
-    "./tasks"
+    "./tasks",
+    "./lib",
+    "lib/config.ts"
   ],
   "files": [
     "./hardhat.config.ts"
